+++ src/leoIntegration.ts
@@ -39,6 +39,7 @@ import { LeoSettingsProvider } from './webviews/leoSettingsWebview';
  */
 export class LeoIntegration {
     // * Status Flags
+    public activated: boolean = true; // Set to false when deactivating the extension
     public finishedStartup: boolean = false;
     private _leoIsConnecting: boolean = false; // Used in connect method, to prevent other attempts while trying
     private _leoBridgeReadyPromise: Promise<LeoBridgePackage> | undefined; // Is set when leoBridge has a leo controller ready
@@ -188,6 +189,13 @@ export class LeoIntegration {
 
     // * Debounced method used to get content of the at-buttons pane
     public refreshButtonsPane: (() => void) & {
+        clear(): void;
+    } & {
+        flush(): void;
+    };
+
+    // * Debounced method used to refresh all
+    public refreshAll: (() => void) & {
         clear(): void;
     } & {
         flush(): void;
@@ -361,7 +369,7 @@ export class LeoIntegration {
             this._onDidOpenTextDocument(p_document)
         );
 
-        // * Debounced refresh flags and UI parts, other than the tree and body, when operation(s) are done executing
+        // * Debounced refresh flags and UI parts, other than the tree and body
         this.getStates = debounce(
             () => { this._triggerGetStates(); },
             Constants.STATES_DEBOUNCE_DELAY
@@ -373,6 +381,18 @@ export class LeoIntegration {
         this.refreshButtonsPane = debounce(
             () => { this._leoButtonsProvider.refreshTreeRoot(); },
             Constants.BUTTONS_DEBOUNCE_DELAY
+        );
+        this.refreshAll = debounce(
+            () => {
+                this.launchRefresh({
+                    tree: true,
+                    body: true,
+                    buttons: true,
+                    states: true,
+                    documents: true
+                }, false);
+            },
+            Constants.REFRESH_ALL_DEBOUNCE_DELAY
         );
     }
 
@@ -461,10 +481,17 @@ export class LeoIntegration {
     }
 
     /**
-     * * Stops the server if it was started by this instance of the extension
+     * * Kills the server process if it was started by this instance of the extension
      */
-    public stopServer(): void {
+    public killServer(): void {
         this._serverService.killServer();
+    }
+
+    /**
+     * * Disconnects from the server
+     */
+    public stopConnection(): void {
+        this._leoBridge.closeLeoProcess();
     }
 
     /**
@@ -477,21 +504,21 @@ export class LeoIntegration {
         }
         this._leoIsConnecting = true;
         console.log('connecting');
-
         this._leoBridgeReadyPromise = this._leoBridge.initLeoProcess(
             this._serverService.usingPort // This will be zero if no port found
         );
         this._leoBridgeReadyPromise.then(
             (p_package) => {
                 console.log('connected');
-
                 this._leoIsConnecting = false;
-                if (p_package.id !== 1) {
+                // Check if hard-coded first package signature
+                if (p_package.id !== Constants.STARTING_PACKAGE_ID) {
                     this.cancelConnect(Constants.USER_MESSAGES.CONNECT_ERROR);
                 } else {
+                    const w_opened: boolean = !!p_package.commander;
                     const w_lastFiles: string[] =
                         this._context.workspaceState.get(Constants.LAST_FILES_KEY) || [];
-                    if (w_lastFiles.length) {
+                    if (w_lastFiles.length && !w_opened) {
                         // This context flag will trigger 'Connecting...' placeholder
                         utils.setContext(Constants.CONTEXT_FLAGS.AUTO_CONNECT, true);
 
@@ -501,6 +528,10 @@ export class LeoIntegration {
                     } else {
                         this.leoStates.leoBridgeReady = true;
                         this.finishedStartup = true;
+                    }
+                    if (w_opened) {
+                        p_package.filename = p_package.commander!.fileName;
+                        this._setupOpenedLeoDocument(p_package);
                     }
 
                     this.showLogPane();
@@ -1140,9 +1171,6 @@ export class LeoIntegration {
             !this._bodyLastChangedDocumentSaved
         ) {
             const w_document = this._bodyLastChangedDocument; // backup for bodySaveDocument before reset
-            if (p_forcedVsCodeSave) {
-                // console.log('FORCED SAVE');
-            }
             this._bodyLastChangedDocumentSaved = true;
             q_savePromise = this._bodySaveDocument(w_document, p_forcedVsCodeSave);
         } else {
@@ -1253,6 +1281,22 @@ export class LeoIntegration {
         } else {
             return Promise.resolve(false);
         }
+    }
+
+    /**
+     * * Sets new body text on leo's side before vscode closes itself if body is dirty
+     * @param p_document Vscode's text document which content will be used to be the new node's body text in Leo
+     * @returns a promise that resolves when the complete saving process is finished
+     */
+    private _bodySaveDeactivate(
+        p_document: vscode.TextDocument
+    ): Promise<LeoBridgePackage> {
+        const w_param = {
+            gnx: utils.leoUriToStr(p_document.uri),
+            body: p_document.getText(),
+        };
+        console.log('Send action set body on quit');
+        return this.sendAction(Constants.LEOBRIDGE.SET_BODY, JSON.stringify(w_param));
     }
 
     /**
@@ -1769,26 +1813,49 @@ export class LeoIntegration {
     /**
      * * cleanupBody closes all remaining body pane to shut down this vscode window
      */
-    public cleanupBody(): Promise<boolean> {
+    public cleanupBody(): Thenable<any> {
+        console.log('start cleanupBody');
+
+        let q_save: Thenable<any>;
+        //
+        if (this._bodyLastChangedDocument &&
+            this._bodyLastChangedDocument.isDirty &&
+            utils.leoUriToStr(this.bodyUri) === utils.leoUriToStr(this._bodyLastChangedDocument.uri)
+        ) {
+            q_save = this._bodySaveDeactivate(this._bodyLastChangedDocument);
+        } else {
+            q_save = Promise.resolve(true);
+        }
+
+        // Adding log in the chain of events
+        q_save.then(() => {
+            console.log('q_save resolved');
+        });
+
         let q_edit: Thenable<boolean>;
         if (this.bodyUri) {
+            console.log('start deleteFile');
+
             const w_edit = new vscode.WorkspaceEdit();
             w_edit.deleteFile(this.bodyUri, { ignoreIfNotExists: true });
-            q_edit = vscode.workspace.applyEdit(w_edit);
+            console.log('applyEdit');
+
+            q_edit = vscode.workspace.applyEdit(w_edit).then(() => {
+                console.log('applyEdit done');
+                return true;
+            });
+        } else {
+            console.log('no deleteFile');
+
+            q_edit = Promise.resolve(true);
         }
-        return Promise.resolve(true)
+        Promise.all([q_save, q_edit])
             .then(() => {
-                if (q_edit) {
-                    console.log('closing body');
-                    return q_edit;
-                } else {
-                    return Promise.resolve(true);
-                }
-            }).then(() => {
-                console.log('done closing body');
+                console.log('cleaned both');
                 return this.closeBody();
             });
 
+        return q_save;
     }
 
     /**
@@ -2420,13 +2487,7 @@ export class LeoIntegration {
             return this.nodeCommand({
                 action: w_action,
                 node: undefined,
-                refreshType: {
-                    tree: true,
-                    body: true,
-                    documents: false,
-                    buttons: false,
-                    states: true,
-                },
+                refreshType: { tree: true, body: true, states: true },
                 fromOutline: false,
             }) || Promise.resolve();
         }